+++ lib/tldr/autorun.rb
@@ -1,0 +1,2 @@
+require "tldr"
+TLDR::Run.at_exit!(TLDR::ArgvParser.new.parse(ARGV))