+++ tests/dotfile_test.rb
@@ -8,20 +8,20 @@ class DotfileTest < Minitest::Test
     assert_includes result.stdout, <<~MSG
       👓
       Command: bundle exec tldr --seed 1 --helper "spec/spec_helper.rb" --no-prepend --base-path "example/c" "spec/math_spec.rb"
-      🌱 --seed 1
+      --seed 1
 
-      🏃 Running:
+      Running:
 
-      😁
+      .
     MSG
   end
 
-  def test_no_dotfile_doesnt_load_those_settings
-    result = TLDRunner.run_command("bundle exec tldr --seed 1 --no-prepend --no-dotfile --base-path example/c")
+  def test_no_config_path_doesnt_load_those_settings
+    result = TLDRunner.run_command("bundle exec tldr --seed 1 --no-prepend --base-path example/c --no-config")
 
     assert_empty result.stderr
     assert_includes result.stdout, <<~MSG
-      Command: bundle exec tldr --seed 1 --no-prepend --base-path "example/c" --no-dotfile
+      Command: bundle exec tldr --seed 1 --no-prepend --base-path "example/c" --no-config
     MSG
     assert_includes result.stdout, "0 test methods"
   end
@@ -31,7 +31,7 @@ class DotfileTest < Minitest::Test
 
     refute result.success?
     assert_includes result.stdout, <<~MSG
-      Command: bundle exec tldr --seed 42 --verbose --helper "test_helper.rb" --load-path "app" --load-path "lib" --parallel --name "/test_*/" --name "test_it" --fail-fast --prepend "a.rb:3" --exclude-path "c.rb:4" --exclude-name "test_b_1" --base-path "example/d" "b.rb"
+      Command: bundle exec tldr --fail-fast --parallel --seed 42 --name "/test_*/" --name "test_it" --exclude-name "test_b_1" --exclude-path "c.rb:4" --helper "test_helper.rb" --prepend "a.rb:3" --load-path "app" --load-path "lib" --base-path "example/d" --verbose "b.rb"
     MSG
     assert_includes result.stderr, <<~MSG
       1) BTest#test_b_2 [b.rb:7] errored:
@@ -47,7 +47,16 @@ class DotfileTest < Minitest::Test
 
     assert result.success?
     assert_includes result.stdout, <<~MSG
-      Command: bundle exec tldr --seed 5 --verbose --helper "test_helper.rb" --load-path "foo" --name "test_stuff" --fail-fast --prepend "nope" --exclude-path "nada" --exclude-name "test_b_2" --base-path "example/d" "b.rb"
+      Command: bundle exec tldr --fail-fast --seed 5 --name "test_stuff" --exclude-name "test_b_2" --exclude-path "nada" --helper "test_helper.rb" --prepend "nope" --load-path "foo" --base-path "example/d" --verbose "b.rb"
+    MSG
+  end
+
+  def test_a_custom_dotfile_path
+    result = TLDRunner.run_command("BUNDLE_GEMFILE=\"example/a/Gemfile\" bundle exec tldr --seed 1 --base-path example/a --config config/TldrFile")
+
+    assert_empty result.stderr
+    assert_includes result.stdout, <<~MSG
+      Command: bundle exec tldr --seed 1 --base-path "example/a" --config config/TldrFile "test/test_subtract.rb"
     MSG
   end
 end