+++ tests/fixture/custom_reporter_helper.rb
@@ -1,0 +1,7 @@
+require_relative "custom_reporter"
+
+module InvalidReporter
+end
+
+class HooklessReporter
+end