+++ tests/line_number_test.rb
@@ -2,7 +2,7 @@ require_relative "test_helper"
 
 class LineNumberTest < Minitest::Test
   def test_line_number_exact_hit
-    result = TLDRunner.should_succeed "line_number.rb:2"
+    result = TLDRunner.should_succeed "line_number.rb:2", "--emoji"
 
     assert_includes result.stdout, "😁"
     refute_includes result.stdout, "🫥"
@@ -10,7 +10,7 @@ class LineNumberTest < Minitest::Test
   end
 
   def test_line_number_intra_method
-    result = TLDRunner.should_succeed "line_number.rb:3"
+    result = TLDRunner.should_succeed "line_number.rb:3", "--emoji"
 
     assert_includes result.stdout, "😁"
     refute_includes result.stdout, "🫥"
@@ -18,7 +18,7 @@ class LineNumberTest < Minitest::Test
   end
 
   def test_line_number_end_of_method
-    result = TLDRunner.should_succeed "line_number.rb:4"
+    result = TLDRunner.should_succeed "line_number.rb:4", "--emoji"
 
     assert_includes result.stdout, "😁"
     refute_includes result.stdout, "🫥"
@@ -26,7 +26,7 @@ class LineNumberTest < Minitest::Test
   end
 
   def test_line_number_two_methods
-    result = TLDRunner.should_fail "line_number.rb:3:11"
+    result = TLDRunner.should_fail "line_number.rb:3:11", "--emoji"
 
     assert_includes result.stdout, "😁"
     assert_includes result.stdout, "😡"
@@ -34,7 +34,7 @@ class LineNumberTest < Minitest::Test
   end
 
   def test_line_number_three_methods
-    result = TLDRunner.should_fail "line_number.rb:3:8:11"
+    result = TLDRunner.should_fail "line_number.rb:3:8:11", "--emoji"
 
     assert_includes result.stdout, "😁"
     assert_includes result.stdout, "🫥"
@@ -42,7 +42,7 @@ class LineNumberTest < Minitest::Test
   end
 
   def test_line_number_three_methods_over_two_patterns
-    result = TLDRunner.should_fail ["line_number.rb:3:11", "line_number.rb:8"]
+    result = TLDRunner.should_fail ["line_number.rb:3:11", "line_number.rb:8"], "--emoji"
 
     assert_includes result.stdout, "😁"
     assert_includes result.stdout, "🫥"
@@ -50,7 +50,7 @@ class LineNumberTest < Minitest::Test
   end
 
   def test_line_number_nonsense
-    result = TLDRunner.should_succeed "line_number.rb:999:42:5"
+    result = TLDRunner.should_succeed "line_number.rb:999:42:5", "--emoji"
 
     refute_includes result.stdout, "😁"
     refute_includes result.stdout, "🫥"