+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## unreleased
+
+* **BREAKING** you know how the whole point of TLDR is that it aborts your test
+run after 1.8s? Yeah, well, it doesn't anymore. Use `--timeout` to enable it
+* **BREAKING** replace the `--no-dotfile` flag and has been replaced by a
+`--[no-]config PATH` flag. To skip loading the YAML file, use `--no-config`.
+To set the file, use `--config FILE` option
+* **BREAKING** Remove `assert_include?` and `refute_include?` in favor of
+Minitest-compatible `assert_includes` and `refute_includes`.
+* **BREAKING** Rename `TLDR::Assertions::MinitestCompatibility` to `TLDR::MinitestCompatibility` and remove `assert_send`, which [nobody uses](https://github.com/minitest/minitest/issues/668)
+* **BREAKING** Replace `no_emoji` YAML option with `emoji` option. Disable emoji output by default. Add `--emoji` flag for enabling it.
+* Add `--[no-]timeout TIMEOUT` flag and `timeout` YAML option. To enable the
+TLDR Classic™ default of 1.8s, specify `--timeout` from the CLI or `timeout: true`
+in YAML. To specify a custom timeout of 42.3 seconds, flag `--timeout 42.3` or
+`timeout: 42.3` in YAML
+* Add `require "tldr/autorun"`, which adds an `at_exit` hook so that tests can be
+run from the command line (still supports CLI args and YAML options) by running `ruby path/to/test.rb` (see [its test](/tests/autorun_test.rb))
+* Fix custom reporters by looking them up only after helpers have a chance to run. [#15](https://github.com/tendersearls/tldr/issues/15)
+
 ## [0.10.1]
 
 * Fix 3.4 / Prism [#17](https://github.com/tendersearls/tldr/pull/17)